requirements.txt
aiohttp
aiosqlite
discord.py==2.6.3
python-dotenv
yt-dlp
Pillow
pillow-heif
pytz
gTTS
ruff